Fixer testing - use specific typed methods over nullable one
