+++ packages/CodingStandard/tests/Fixer/Commenting/BlockPropertyCommentFixer/BlockPropertyCommentFixerTest.php
@@ -10,36 +10,29 @@ use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 final class BlockPropertyCommentFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
-     * @dataProvider provideFixCases()
+     * @dataProvider provideWrongToFixedCases()
      */
-    public function testFix(string $expected, string $input): void
+    public function testWrongToFixedCases(string $wrongFile, string $correctFile): void
     {
-        $this->doTest($expected, $input);
+        $this->doTestWrongToFixedFile($wrongFile, $correctFile);
     }
 
     /**
      * @return string[][]
      */
-    public function provideFixCases(): array
+    public function provideWrongToFixedCases(): array
     {
         return [
-            # wrong => fixed
-            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
-            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
+            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc'],
+            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc'],
         ];
     }
 
     protected function createFixer(): FixerInterface
     {
-        $blockPropertyCommentFixer = new BlockPropertyCommentFixer();
-        $blockPropertyCommentFixer->setWhitespacesConfig($this->createWhitespacesFixerConfig());
+        $fixer = new BlockPropertyCommentFixer();
+        $fixer->setWhitespacesConfig(new WhitespacesFixerConfig('    ', PHP_EOL));
 
-        return $blockPropertyCommentFixer;
-    }
-
-    private function createWhitespacesFixerConfig(): WhitespacesFixerConfig
-    {
-        // @try tabs?
-        return new WhitespacesFixerConfig('    ', PHP_EOL);
+        return $fixer;
     }
 }
