+++ packages/CodingStandard/tests/Fixer/Naming/ExceptionNameFixer/ExceptionNameFixerTest.php
@@ -9,20 +9,19 @@ use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 final class ExceptionNameFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
-     * @dataProvider provideFixCases()
+     * @dataProvider provideWrongToFixedCases()
      */
-    public function testFix(string $expected, string $input): void
+    public function testWrongToFixed(string $wrongFile, string $fixedFile): void
     {
-        $this->doTest($expected, $input);
+        $this->doTestWrongToFixedFile($wrongFile, $fixedFile);
     }
 
     /**
      * @return string[][]
      */
-    public function provideFixCases(): array
+    public function provideWrongToFixedCases(): array
     {
         return [
-            # wrong => fixed
             [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc'],
         ];
     }
