+++ packages/CodingStandard/tests/Fixer/Php/ClassStringToClassConstantFixer/ClassStringToClassConstantFixerTest.php
@@ -9,27 +9,40 @@ use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 final class ClassStringToClassConstantFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
-     * @dataProvider provideFixCases()
+     * @dataProvider provideCorrectCases()
      */
-    public function testFix(string $expected, ?string $input = null): void
+    public function testCorrectCases(string $file): void
     {
-        $this->doTest($expected, $input);
+        $this->doTestCorrectFile($file);
     }
 
     /**
      * @return string[][]
      */
-    public function provideFixCases(): array
+    public function provideCorrectCases(): array
     {
         return [
-            # wrong => fixed
-            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
-            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
-            [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc', ],
-            # correct
-            ['<?php $form->addText(\'datetime\');'],
-            ['<?php $request->getParameter(\'exception\');'],
-            ['<?php $this->assertTrue(class_exists(\'\ApiGen\Reflection\Tests\ExtendingClass\'));'],
+            [__DIR__ . '/correct/correct.php.inc'],
+        ];
+    }
+
+    /**
+     * @dataProvider provideWrongToFixedCases()
+     */
+    public function testWrongToFixed(string $wrongFile, string $fixedFile): void
+    {
+        $this->doTestWrongToFixedFile($wrongFile, $fixedFile);
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function provideWrongToFixedCases(): array
+    {
+        return [
+            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc'],
+            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc'],
+            [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc'],
         ];
     }
