+++ packages/CodingStandard/tests/Fixer/Php/ClassStringToClassConstantFixer/correct/correct.php.inc
@@ -1,0 +1,7 @@
+<?php
+
+$form->addText('datetime');
+
+$request->getParameter('exception');
+
+$this->assertTrue(class_exists('\ApiGen\Reflection\Tests\ExtendingClass'));
