+++ packages/CodingStandard/tests/Fixer/Property/ArrayPropertyDefaultValueFixer/ArrayPropertyDefaultValueFixerTest.php
@@ -9,26 +9,42 @@ use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 final class ArrayPropertyDefaultValueFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
-     * @dataProvider provideFixCases()
+     * @dataProvider provideCorrectCases()
      */
-    public function testFix(string $expected, ?string $input = null): void
+    public function testCorrectCases(string $file): void
     {
-        $this->doTest($expected, $input);
+        $this->doTestCorrectFile($file);
     }
 
     /**
      * @return string[][]
      */
-    public function provideFixCases(): array
+    public function provideCorrectCases(): array
     {
         return [
-            # wrong => fixed
-            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
-            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
-            [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc', ],
-            [__DIR__ . '/wrong/wrong4.php.inc', __DIR__ . '/fixed/fixed4.php.inc', ],
-            # correct
-            [__DIR__ . '/correct/correct.php.inc', ], ];
+            [__DIR__ . '/correct/correct.php.inc'],
+        ];
+    }
+
+    /**
+     * @dataProvider provideWrongToFixedCases()
+     */
+    public function testWrongToFixed(string $wrongFile, string $fixedFile): void
+    {
+        $this->doTestWrongToFixedFile($wrongFile, $fixedFile);
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function provideWrongToFixedCases(): array
+    {
+        return [
+            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc'],
+            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc'],
+            [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc'],
+            [__DIR__ . '/wrong/wrong4.php.inc', __DIR__ . '/fixed/fixed4.php.inc'],
+        ];
     }
 
     protected function createFixer(): FixerInterface
