+++ packages/CodingStandard/tests/Fixer/Solid/FinalInterfaceFixer/ConfiguredTest.php
@@ -9,20 +9,19 @@ use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 final class ConfiguredTest extends AbstractSimpleFixerTestCase
 {
     /**
-     * @dataProvider provideFixCases()
+     * @dataProvider provideWrongToFixedCases()
      */
-    public function testFix(string $expected, ?string $input = null): void
+    public function testWrongToFixed(string $wrongFile, string $fixedFile): void
     {
-        $this->doTest($expected, $input);
+        $this->doTestWrongToFixedFile($wrongFile, $fixedFile);
     }
 
     /**
      * @return string[][]
      */
-    public function provideFixCases(): array
+    public function provideWrongToFixedCases(): array
     {
         return [
-            # wrong => fixed
             [__DIR__ . '/wrong/wrong3.php.inc', __DIR__ . '/fixed/fixed3.php.inc'],
         ];
     }
@@ -32,7 +31,7 @@ final class ConfiguredTest extends AbstractSimpleFixerTestCase
         $fixer = new FinalInterfaceFixer();
 
         $fixer->configure([
-            'only_interfaces' => ['SomeInterface'],
+            FinalInterfaceFixer::ONLY_INTERFACES_OPTION => ['SomeInterface'],
         ]);
 
         return $fixer;
