+++ packages/CodingStandard/tests/Fixer/Solid/FinalInterfaceFixer/FinalInterfaceFixerTest.php
@@ -9,28 +9,43 @@ use Symplify\TokenRunner\Testing\AbstractSimpleFixerTestCase;
 final class FinalInterfaceFixerTest extends AbstractSimpleFixerTestCase
 {
     /**
-     * @dataProvider provideFixCases()
+     * @dataProvider provideCorrectCases()
      */
-    public function testFix(string $expected, ?string $input = null): void
+    public function testCorrectCases(string $file): void
     {
-        $this->doTest($expected, $input);
+        $this->doTestCorrectFile($file);
     }
 
     /**
      * @return string[][]
      */
-    public function provideFixCases(): array
+    public function provideCorrectCases(): array
     {
         return [
-            # wrong => fixed
-            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc', ],
-            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc', ],
-            # correct
             [__DIR__ . '/correct/correct.php.inc'],
             [__DIR__ . '/correct/correct2.php.inc'],
             [__DIR__ . '/correct/correct3.php.inc'],
             [__DIR__ . '/correct/correct4.php.inc'],
             [__DIR__ . '/correct/correct5.php.inc'],
+        ];
+    }
+
+    /**
+     * @dataProvider provideWrongToFixedCases()
+     */
+    public function testWrongToFixed(string $wrongFile, string $fixedFile): void
+    {
+        $this->doTestWrongToFixedFile($wrongFile, $fixedFile);
+    }
+
+    /**
+     * @return string[][]
+     */
+    public function provideWrongToFixedCases(): array
+    {
+        return [
+            [__DIR__ . '/wrong/wrong.php.inc', __DIR__ . '/fixed/fixed.php.inc'],
+            [__DIR__ . '/wrong/wrong2.php.inc', __DIR__ . '/fixed/fixed2.php.inc'],
         ];
     }
