+++ packages/CodingStandard/tests/Fixer/Strict/BlankLineAfterStrictTypesFixer/correct/correct.php.inc
@@ -1,0 +1,1 @@
+<?php declare(strict_types=1);
