+++ packages/CodingStandard/tests/Fixer/Strict/BlankLineAfterStrictTypesFixer/fixed/fixed.php.inc
@@ -1,0 +1,3 @@
+<?php declare(strict_types=1);
+
+namespace SomeNamespace;
