+++ packages/CodingStandard/tests/Fixer/Strict/BlankLineAfterStrictTypesFixer/wrong/wrong.php.inc
@@ -1,0 +1,2 @@
+<?php declare(strict_types=1);
+namespace SomeNamespace;
