+++ packages/CodingStandard/tests/Fixer/Strict/BlankLineAfterStrictTypesFixer/wrong/wrong2.php.inc
@@ -1,0 +1,4 @@
+<?php declare(strict_types=1);
+
+
+namespace SomeNamespace;
